solucionado excepcion

+++ practica2/clases/pedido.hpp
@@ -18,14 +18,14 @@ class Pedido{
         int beneficioPedido(){return beneficio;}
         double beneficioIndividual(){return (double)this->beneficio/this->numBilletes;}
         bool operator >=(Pedido& p) {
-            if(this->beneficio/this->numBilletes == p.beneficioIndividual()){
+            if((double)this->beneficio/this->numBilletes == p.beneficioIndividual()){
                 return numBilletes >= p.billetes();
             }
 			return (double)this->beneficio/this->numBilletes >= p.beneficioIndividual();
 		}
-        
+
 		bool operator <=(Pedido& p) {
-            if(this->beneficio/this->numBilletes == p.beneficioIndividual()){
+            if((double)this->beneficio/this->numBilletes == p.beneficioIndividual()){
                 return numBilletes <= p.billetes();
             }
 			return (double)this->beneficio/this->numBilletes <= p.beneficioIndividual();		
@@ -33,7 +33,7 @@ class Pedido{
 
 
 		bool operator >(Pedido& p) {
-            if(this->beneficio/this->numBilletes == p.beneficioIndividual()){
+            if((double)this->beneficio/this->numBilletes == p.beneficioIndividual()){
                 return numBilletes > p.billetes();
             }
 			return (double)this->beneficio/this->numBilletes > p.beneficioIndividual();		
@@ -42,7 +42,7 @@ class Pedido{
 
 
 		bool operator <(Pedido& p) {
-            if(this->beneficio/this->numBilletes == p.beneficioIndividual()){
+            if((double)this->beneficio/this->numBilletes == p.beneficioIndividual()){
                 return numBilletes < p.billetes();
             }
 			return (double)this->beneficio/this->numBilletes < p.beneficioIndividual();
